requirements.txt
fastapi
uvicorn
transformers
torch
pydantic
scikit-learn
python-dotenv
requests
groq
google-generativeai